@@ -12,53 +12,55 @@
         ]
     },
     "dependencies": {
-        "@react-native-community/async-storage": "1.11.0",
+        "@react-native-community/async-storage": "1.12.0",
         "@react-native-community/masked-view": "0.1.10",
-        "@react-native-community/netinfo": "5.9.5",
-        "@react-navigation/native": "5.0.9",
-        "@react-navigation/stack": "5.1.1",
-        "@teamberry/muscadine": "^0.7.0",
+        "@react-native-community/netinfo": "5.9.6",
+        "@react-navigation/native": "5.7.3",
+        "@react-navigation/stack": "5.9.0",
+        "@teamberry/muscadine": "^0.8.4",
         "axios": "0.19.2",
-        "expo": "36.0.2",
+        "expo": "38.0.9",
         "formik": "2.1.5",
-        "react": "16.9.0",
-        "react-dom": "16.9.0",
-        "react-native": "0.61.5",
+        "react": "16.13.1",
+        "react-dom": "16.13.1",
+        "react-native": "0.63.2",
         "react-native-collapsible": "1.5.3",
+        "react-native-config": "1.3.3",
         "react-native-gesture-handler": "1.7.0",
-        "react-native-paper": "3.11.0",
-        "react-native-reanimated": "1.10.1",
-        "react-native-safe-area-context": "0.7.3",
-        "react-native-screens": "2.9.0",
+        "react-native-paper": "4.0.1",
+        "react-native-reanimated": "1.10.2",
+        "react-native-safe-area-context": "3.1.4",
+        "react-native-screens": "2.10.1",
         "react-native-side-menu": "1.1.3",
         "react-native-svg": "12.1.0",
-        "react-native-tab-view": "2.15.0",
-        "react-native-unimodules": "0.7.0",
-        "react-native-vector-icons": "6.6.0",
-        "react-native-web": "0.11.7",
-        "react-native-youtube": "2.0.0",
+        "react-native-swipe-list-view": "3.2.3",
+        "react-native-tab-view": "2.15.1",
+        "react-native-unimodules": "0.10.1",
+        "react-native-vector-icons": "7.0.0",
+        "react-native-web": "0.13.8",
+        "react-native-youtube": "2.0.1",
         "socket.io-client": "2.3.0",
-        "yup": "0.28.5"
+        "yup": "0.29.3"
     },
     "devDependencies": {
-        "@babel/core": "7.10.5",
-        "@types/react": "16.9.43",
-        "@types/react-native": "0.63.2",
-        "@types/yup": "0.28.1",
-        "@typescript-eslint/eslint-plugin": "2.34.0",
-        "@typescript-eslint/parser": "2.34.0",
-        "babel-jest": "24.9.0",
+        "@babel/core": "7.11.1",
+        "@types/react": "16.9.46",
+        "@types/react-native": "0.63.8",
+        "@types/yup": "0.29.5",
+        "@typescript-eslint/eslint-plugin": "3.9.1",
+        "@typescript-eslint/parser": "3.9.1",
+        "babel-jest": "26.3.0",
         "babel-preset-expo": "8.2.3",
-        "eslint": "6.8.0",
+        "eslint": "7.7.0",
         "eslint-config-airbnb": "18.2.0",
         "eslint-plugin-import": "2.22.0",
         "eslint-plugin-jsx-a11y": "6.3.1",
-        "eslint-plugin-react": "7.20.5",
-        "eslint-plugin-react-hooks": "2.5.1",
-        "jest": "24.9.0",
-        "metro-react-native-babel-preset": "0.61.0",
+        "eslint-plugin-react": "7.20.6",
+        "eslint-plugin-react-hooks": "4.1.0",
+        "jest": "26.4.0",
+        "metro-react-native-babel-preset": "0.62.0",
         "react-native-svg-transformer": "0.14.3",
-        "react-test-renderer": "16.9.0",
+        "react-test-renderer": "16.13.1",
         "typescript": "3.9.7"
     },
     "private": true,
