@@ -7,6 +7,7 @@ import Box from '../models/box.model';
 import ProfilePicture from './profile-picture.component';
 
 import UsersIcon from '../../assets/icons/users-icon.svg';
+import BxChipComponent from './bx-chip.component';
 
 const styles = StyleSheet.create({
   card: {
@@ -49,17 +50,6 @@ const styles = StyleSheet.create({
     flex: 1,
     flexDirection: 'row',
   },
-  boxMode: {
-    backgroundColor: '#979797',
-    color: 'white',
-    borderRadius: 10,
-    width: 80,
-    height: 18,
-    textAlign: 'center',
-    marginRight: 3,
-    marginLeft: 3,
-    fontSize: 12,
-  },
   boxUserDisplay: {
     backgroundColor: 'rgba(0, 0, 0, 0.8)',
     borderRadius: 5,
@@ -75,7 +65,8 @@ const styles = StyleSheet.create({
 
 const displayCurrentvideo = (box: Box): QueueItem => box.playlist.find((video) => video.startTime !== null && video.endTime === null);
 
-const BoxCard = (box: Box) => {
+const BoxCard = (props: { box: Box }) => {
+  const { box } = props;
   const currentVideo = displayCurrentvideo(box);
 
   return (
@@ -107,18 +98,32 @@ const BoxCard = (box: Box) => {
           <Text style={styles.boxCurrent} numberOfLines={2}>{currentVideo ? currentVideo.video.name : '--'}</Text>
         </View>
       </View>
-      <View style={styles.boxMode}>
-        <View style={styles.boxModes}>
-          {box.options.random ? (
-            <Text style={styles.boxMode}>Random</Text>
-          ) : (<></>)}
-          {box.options.loop ? (
-            <Text style={styles.boxMode}>Loop</Text>
-          ) : (<></>)}
-          {!box.options.berries ? (
-            <Text style={styles.boxMode}>No Berries</Text>
-          ) : (<></>)}
-        </View>
+      <View style={styles.boxModes}>
+        {box.options.random ? (
+          <View style={{ paddingHorizontal: 2 }}>
+            <BxChipComponent options={{ type: 'random', chipText: 'Random' }} display="full" />
+          </View>
+        ) : null}
+        {box.options.loop ? (
+          <View style={{ paddingHorizontal: 2 }}>
+            <BxChipComponent options={{ type: 'loop', chipText: 'Loop' }} display="full" />
+          </View>
+        ) : null}
+        {box.options.berries ? (
+          <View style={{ paddingHorizontal: 2 }}>
+            <BxChipComponent options={{ type: 'coin-enabled', chipText: 'Berries' }} display="full" />
+          </View>
+        ) : null}
+        {box.private ? (
+          <View style={{ paddingHorizontal: 2 }}>
+            <BxChipComponent options={{ type: 'lock', chipText: 'Private' }} display="full" />
+          </View>
+        ) : null}
+        {box.options.videoMaxDurationLimit !== 0 ? (
+          <View style={{ paddingHorizontal: 2 }}>
+            <BxChipComponent options={{ type: 'duration-limit', chipText: `${box.options.videoMaxDurationLimit} mins` }} display="full" />
+          </View>
+        ) : null}
       </View>
     </View>
   );
