@@ -2,32 +2,37 @@ import React, { useState } from 'react';
 import { Dimensions } from 'react-native';
 import { TabView, TabBar } from 'react-native-tab-view';
 
+import { Permission } from '@teamberry/muscadine';
 import ChatTab from './chat-tab.component';
 import SearchTab from './search-tab.component';
 import Box from '../../../models/box.model';
 
-const Panel = (props: { box: Box, socket: any}) => {
+const Panel = (props: { box: Box, socket: any, berryCount: number, permissions: Array<Permission> }) => {
   const [index, setIndex] = useState(0);
   const [routes] = useState([
     { key: 'chat', title: 'Chat' },
     { key: 'search', title: 'Search' },
   ]);
+  const {
+    box, socket, berryCount, permissions,
+  } = props;
 
   const initialLayout = { width: Dimensions.get('window').width };
 
   const renderScene = ({ route }) => {
     switch (route.key) {
       case 'chat':
-        return <ChatTab boxToken={props.box._id} socket={props.socket} />;
+        return <ChatTab box={box} socket={socket} berryCount={berryCount} />;
       case 'search':
-        return <SearchTab box={props.box} socket={props.socket} />;
+        return <SearchTab box={box} socket={socket} berryCount={berryCount} permissions={permissions} />;
       default:
-        return <></>;
+        return null;
     }
   };
 
   const renderTabBar = (props) => (
     <TabBar
+      // eslint-disable-next-line react/jsx-props-no-spreading
       {...props}
       style={{ backgroundColor: '#191919' }}
       indicatorStyle={{ backgroundColor: '#009AEB' }}
