@@ -1,23 +1,33 @@
-import React, { useState, useRef } from 'react';
+import React, { useState, useRef, useEffect } from 'react';
 import {
-  StyleSheet, Text, View, FlatList, TouchableOpacity, Animated,
+  StyleSheet, Text, View, FlatList, TouchableOpacity, Animated, Pressable, TextInput,
 } from 'react-native';
-import { QueueItem } from '@teamberry/muscadine';
+import { QueueItem, Permission } from '@teamberry/muscadine';
 import Collapsible from 'react-native-collapsible';
 import { Svg, Polygon } from 'react-native-svg';
+import axios from 'axios';
+import Config from 'react-native-config';
+import { Snackbar } from 'react-native-paper';
 import Box from '../../../models/box.model';
 import QueueVideo from './queue-video.component';
 import ProfilePicture from '../../../components/profile-picture.component';
 
+import RandomIcon from '../../../../assets/icons/random-icon.svg';
+import ReplayIcon from '../../../../assets/icons/replay-icon.svg';
+import BerriesIcon from '../../../../assets/icons/berry-coin-icon.svg';
+import BerriesEnabledIcon from '../../../../assets/icons/coin-enabled-icon.svg';
+import DurationRestrictionIcon from '../../../../assets/icons/duration-limit-icon.svg';
+import BerryCounter from './berry-counter.component';
+
 const styles = StyleSheet.create({
   currentSpaceContainer: {
     height: 50,
-    backgroundColor: '#262626',
+    backgroundColor: '#111111',
     color: 'white',
     paddingLeft: 10,
   },
   currentSpace: {
-    flex: 1,
+    display: 'flex',
     flexDirection: 'row',
     alignItems: 'center',
     justifyContent: 'space-between',
@@ -35,29 +45,78 @@ const styles = StyleSheet.create({
   },
 });
 
-const Queue = ({ box, currentVideo }: {
+const Queue = (props: {
     box: Box,
-    currentVideo: QueueItem
+    currentVideo: QueueItem,
+    height: number,
+    berryCount: number,
+    permissions: Array<Permission>
 }) => {
-  const [isCollapsed, setCollapse] = useState(true);
+  const {
+    box, currentVideo, height, berryCount, permissions,
+  } = props;
 
-  const BoxName = () => {
-    if (box) {
-      return (
-        <Text style={styles.boxName}>{box.name}</Text>
-      );
+  const _durationInputRef = useRef(null);
+
+  const [isCollapsed, setCollapse] = useState(true);
+  const [error, setError] = useState(false);
+  const [hasUpdatedSuccessfully, setUpdateState] = useState(false);
+  const [isDurationInputVisible, setDurationInputVisibility] = useState(false);
+  const [queueVideos, setQueueVideos] = useState([] as Array<QueueItem>);
+
+  useEffect((): void => {
+    if (!box) {
+      setQueueVideos([]);
     }
-    return (<></>);
-  };
+
+    let upcomingVideos = [];
+
+    if (!box.options.loop) {
+      upcomingVideos = box.playlist.filter((item) => item.startTime === null);
+    } else {
+      upcomingVideos = box.playlist;
+    }
+
+    upcomingVideos.reverse();
+
+    // Put the preslected video first
+    const preselectedVideoIndex = upcomingVideos.findIndex((item: QueueItem) => item.isPreselected);
+    if (preselectedVideoIndex !== -1) {
+      const preselectedVideo = upcomingVideos[preselectedVideoIndex];
+      upcomingVideos.splice(preselectedVideoIndex, 1);
+      upcomingVideos.unshift(preselectedVideo);
+    }
+
+    // Put the current video first
+    const playingVideo: QueueItem = box.playlist.find((item: QueueItem) => item.startTime !== null && item.endTime === null);
+    if (playingVideo) {
+      // If loop, the full queue is displayed, regardless of the state of the videos.
+      // So the current video has to be spliced out before being unshifted.
+      const playingVideoIndex = upcomingVideos.findIndex((item: QueueItem) => item.video.link === playingVideo.video.link);
+      if (playingVideoIndex !== -1) {
+        upcomingVideos.splice(playingVideoIndex, 1);
+      }
+      upcomingVideos.unshift(playingVideo);
+    }
+
+    setQueueVideos(upcomingVideos);
+  }, [box.playlist]);
 
   const CurrentVideo = () => {
-    if (currentVideo) {
-      return (
-        <Text style={styles.currentVideo} numberOfLines={1}>{currentVideo.video.name}</Text>
-      );
+    if (!currentVideo) {
+      return null;
     }
 
-    return (<></>);
+    return (
+      <View style={{ flex: 0, flexDirection: 'row' }}>
+        {currentVideo.stateForcedWithBerries ? (
+          <View style={{ paddingRight: 5 }}>
+            <BerriesIcon width={20} height={20} />
+          </View>
+        ) : null}
+        <Text style={styles.currentVideo} numberOfLines={1}>{currentVideo.video.name}</Text>
+      </View>
+    );
   };
 
   const rotateAnim = useRef(new Animated.Value(0)).current;
@@ -96,58 +155,119 @@ const Queue = ({ box, currentVideo }: {
     setCollapse(!isCollapsed);
   };
 
-  const QueueList = () => {
-    if (!box) {
-      return (<></>);
+  const patchBox = async (setting) => {
+    try {
+      await axios.patch(`${Config.API_URL}/boxes/${box._id}`, setting);
+      setUpdateState(true);
+    } catch (error) {
+      setError(true);
     }
+  };
 
-    let upcomingVideos = [];
-
-    if (!box.options.loop) {
-      upcomingVideos = box.playlist.filter((item) => item.startTime === null);
-    } else {
-      upcomingVideos = box.playlist;
-    }
-
-    if (upcomingVideos.length === 0) {
+  const RandomIndicator = () => {
+    if (permissions.includes('editBox')) {
       return (
-        <Text style={{ textAlign: 'center', color: '#BBB', marginHorizontal: 20 }}>The Queue is empty.</Text>
+        <Pressable onPress={() => { patchBox({ random: !box.options.random }); }}>
+          <RandomIcon width={20} height={20} fill={box.options.random ? '#009AEB' : '#CCCCCC'} />
+        </Pressable>
       );
     }
 
-    upcomingVideos.reverse();
-
-    // Put the preslected video first
-    const preselectedVideoIndex = upcomingVideos.findIndex((item: QueueItem) => item.isPreselected);
-    if (preselectedVideoIndex !== -1) {
-      const preselectedVideo = upcomingVideos[preselectedVideoIndex];
-      upcomingVideos.splice(preselectedVideoIndex, 1);
-      upcomingVideos.unshift(preselectedVideo);
+    if (box.options.random) {
+      return <RandomIcon width={20} height={20} fill="#009AEB" />;
     }
 
-    // Put the current video first
-    const playingVideo: QueueItem = box.playlist.find((item: QueueItem) => item.startTime !== null && item.endTime === null);
-    if (playingVideo) {
-      // If loop, the full queue is displayed, regardless of the state of the videos.
-      // So the current video has to be spliced out before being unshifted.
-      const playingVideoIndex = upcomingVideos.findIndex((item: QueueItem) => item.video.link === playingVideo.video.link);
-      if (playingVideoIndex !== -1) {
-        upcomingVideos.splice(playingVideoIndex, 1);
-      }
-      upcomingVideos.unshift(playingVideo);
-    }
-
-    return (
-      <FlatList
-        data={upcomingVideos}
-        ItemSeparatorComponent={() => <View style={{ backgroundColor: '#191919', height: 1 }} />}
-        renderItem={({ item }) => (
-          <QueueVideo item={item} />
-        )}
-        keyExtractor={(item, index) => index.toString()}
-      />
-    );
+    return null;
   };
+
+  const LoopIndicator = () => {
+    if (permissions.includes('editBox')) {
+      return (
+        <Pressable onPress={() => { patchBox({ loop: !box.options.loop }); }}>
+          <ReplayIcon width={20} height={20} fill={box.options.loop ? '#009AEB' : '#CCCCCC'} />
+        </Pressable>
+      );
+    }
+
+    if (box.options.loop) {
+      return <ReplayIcon width={20} height={20} fill="#009AEB" />;
+    }
+
+    return null;
+  };
+
+  const BerriesIndicator = () => {
+    if (permissions.includes('editBox')) {
+      return (
+        <Pressable onPress={() => { patchBox({ berries: !box.options.berries }); }}>
+          <BerriesEnabledIcon width={20} height={20} fill={box.options.berries ? '#009AEB' : '#CCCCCC'} />
+        </Pressable>
+      );
+    }
+
+    if (box.options.berries) {
+      return <BerryCounter count={berryCount} />;
+    }
+
+    return null;
+  };
+
+  const DurationRestrictionIndicator = () => {
+    if (permissions.includes('editBox')) {
+      if (box.options.videoMaxDurationLimit !== 0) {
+        return (
+          <Pressable onPress={() => { patchBox({ videoMaxDurationLimit: 0 }); }}>
+            <View style={{ flex: 0, flexDirection: 'row' }}>
+              <DurationRestrictionIcon width={20} height={20} fill={box.options.videoMaxDurationLimit !== 0 ? '#009AEB' : '#CCCCCC'} />
+              {box.options.videoMaxDurationLimit ? (
+                <Text style={{ color: '#009AEB' }}>
+                  {box.options.videoMaxDurationLimit}
+                  {' '}
+                  mins
+                </Text>
+              ) : null}
+            </View>
+          </Pressable>
+        );
+      }
+      return (
+        <Pressable onPress={() => { setDurationInputVisibility(!isDurationInputVisible); }}>
+          <View style={{ flex: 0, flexDirection: 'row' }}>
+            <DurationRestrictionIcon width={20} height={20} fill="#CCCCCC" />
+          </View>
+        </Pressable>
+      );
+    }
+
+    if (box.options.videoMaxDurationLimit) {
+      return (
+        <View style={{ flex: 0, flexDirection: 'row' }}>
+          <DurationRestrictionIcon width={20} height={20} fill={box.options.videoMaxDurationLimit !== 0 ? '#009AEB' : '#CCCCCC'} />
+          <Text style={{ color: '#009AEB' }}>
+            {box.options.videoMaxDurationLimit}
+            {' '}
+            mins
+          </Text>
+        </View>
+      );
+    }
+
+    return null;
+  };
+
+  const QueueList = () => (
+    <FlatList
+      data={queueVideos}
+      ItemSeparatorComponent={() => <View style={{ backgroundColor: '#191919', height: 1 }} />}
+      renderItem={({ item }) => (
+        <QueueVideo item={item} boxToken={box._id} permissions={permissions} berriesEnabled={box.options.berries} />
+      )}
+      keyExtractor={(item) => item._id}
+      initialNumToRender={8}
+      windowSize={12}
+      ListEmptyComponent={() => <Text style={{ textAlign: 'center', color: '#BBB', marginHorizontal: 20 }}>The Queue is empty.</Text>}
+    />
+  );
 
   return (
     <>
@@ -157,11 +277,11 @@ const Queue = ({ box, currentVideo }: {
       >
         <View style={styles.currentSpaceContainer}>
           <View style={styles.currentSpace}>
-            <View style={{ width: 35 }}>
+            <View style={{ paddingRight: 10 }}>
               <ProfilePicture userId={box.creator._id} size={25} />
             </View>
-            <View style={{ width: 275 }}>
-              <BoxName />
+            <View style={{ flex: 1 }}>
+              <Text style={styles.boxName}>{box.name}</Text>
               <CurrentVideo />
             </View>
             <View style={{ width: 40 }}>
@@ -180,10 +300,78 @@ const Queue = ({ box, currentVideo }: {
         </View>
       </TouchableOpacity>
       <View style={styles.upcomingSpaceContainer}>
-        <Collapsible collapsed={isCollapsed} style={{ height: 445 }}>
+        <Collapsible
+          collapsed={isCollapsed}
+          style={{ height: height - 50 }}
+        >
+          <View style={{ backgroundColor: '#191919', padding: 10 }}>
+            <View style={{ flex: 0, flexDirection: 'row', justifyContent: 'space-between' }}>
+              {/* Display icons in "inactive" status for users who can act on them. Don't display icons for user who cannot. */}
+              <View style={{ flex: 0, flexDirection: 'row' }}>
+                <View style={{ paddingRight: 20 }}>
+                  <RandomIndicator />
+                </View>
+                <View style={{ paddingRight: 20 }}>
+                  <LoopIndicator />
+                </View>
+                <View style={{ paddingRight: 20 }}>
+                  <BerriesIndicator />
+                </View>
+                <View style={{ paddingRight: 20 }}>
+                  <DurationRestrictionIndicator />
+                </View>
+              </View>
+            </View>
+            {isDurationInputVisible ? (
+              <View style={{ paddingVertical: 5 }}>
+                <TextInput
+                  ref={_durationInputRef}
+                  keyboardType="numeric"
+                  onSubmitEditing={({ nativeEvent }) => { patchBox({ videoMaxDurationLimit: nativeEvent.text }); }}
+                  onBlur={() => setDurationInputVisibility(false)}
+                  placeholder="Set the duration restriction (in minutes)"
+                  autoFocus
+                  placeholderTextColor="#CCCCCC"
+                  style={{
+                    height: 40,
+                    borderWidth: 1,
+                    borderStyle: 'solid',
+                    borderColor: '#009aeb',
+                    padding: 10,
+                    borderRadius: 5,
+                    color: 'white',
+                  }}
+                />
+              </View>
+            ) : null}
+          </View>
           <QueueList />
         </Collapsible>
       </View>
+      <Snackbar
+        visible={hasUpdatedSuccessfully}
+        duration={1500}
+        style={{
+          backgroundColor: '#090909',
+          borderLeftColor: '#0CEBC0',
+          borderLeftWidth: 10,
+        }}
+        onDismiss={() => setUpdateState(false)}
+      >
+        Option updated successfully.
+      </Snackbar>
+      <Snackbar
+        visible={error}
+        duration={5000}
+        style={{
+          backgroundColor: '#090909',
+          borderLeftColor: '#B30F4F',
+          borderLeftWidth: 10,
+        }}
+        onDismiss={() => setError(false)}
+      >
+        Something wrong happened. Try again?
+      </Snackbar>
     </>
   );
 };
