@@ -1,14 +1,18 @@
 import React, { useState, useEffect } from 'react';
 import {
-  Image, View, Text, StyleSheet, TextInput, FlatList, ActivityIndicator,
+  Image, View, Text, StyleSheet, TextInput, FlatList, Pressable,
 } from 'react-native';
 import axios from 'axios';
 import AsyncStorage from '@react-native-community/async-storage';
-import { VideoSubmissionRequest, QueueItem } from '@teamberry/muscadine';
-import { TouchableWithoutFeedback } from 'react-native-gesture-handler';
+import { VideoSubmissionRequest, QueueItem, Permission } from '@teamberry/muscadine';
 import { Snackbar } from 'react-native-paper';
+import Config from 'react-native-config';
+
 import Box from '../../../models/box.model';
 import DurationIndicator from '../../../components/duration-indicator.component';
+import BxLoadingIndicator from '../../../components/bx-loading-indicator.component';
+import BxButtonComponent from '../../../components/bx-button.component';
+import BerryCounter from './berry-counter.component';
 
 const styles = StyleSheet.create({
   container: {
@@ -29,6 +33,7 @@ const styles = StyleSheet.create({
     borderWidth: 1,
     borderRadius: 5,
     color: 'white',
+    flex: 1,
   },
   resultsHelp: {
     color: '#BBBBBB',
@@ -39,8 +44,7 @@ const styles = StyleSheet.create({
   resultItem: {
     paddingHorizontal: 7,
     paddingVertical: 10,
-    flex: 1,
-    flexDirection: 'row',
+    flexDirection: 'column',
   },
   inQueueIndicator: {
     fontFamily: 'Montserrat-SemiBold',
@@ -52,17 +56,27 @@ const styles = StyleSheet.create({
   },
 });
 
-const SearchTab = (props: {socket: any, box: Box}) => {
+export interface Video {
+    _id?: string,
+    name: string,
+    link: string;
+    duration?: string;
+}
+
+const SearchTab = (props: {socket: any, box: Box, berryCount: number, permissions: Array<Permission>}) => {
+  const {
+    socket, box, berryCount, permissions,
+  } = props;
+
   const [searchValue, setSearchValue] = useState('');
-  const [youtubeSearchResults, setSearchResults] = useState([]);
+  const [youtubeSearchResults, setSearchResults] = useState([] as Array<Video>);
   const [user, setUser] = useState(null);
   const [hasSearched, setSearched] = useState(false);
   const [isSearching, setSearching] = useState(false);
   const [error, setError] = useState(false);
   const [isSubmitted, setSubmitted] = useState(false);
   const [videosInQueue, setQueueIds] = useState([]);
-
-  const { socket, box } = props;
+  const [boxOptions, setBoxOptions] = useState(box.options);
 
   useEffect(() => {
     const getSession = async () => {
@@ -78,6 +92,7 @@ const SearchTab = (props: {socket: any, box: Box}) => {
     setQueueIds(videoIds);
 
     socket.on('box', (box: Box) => {
+      setBoxOptions(box.options);
       const videoIds = box.playlist.map((queueItem: QueueItem) => queueItem.video.link);
       setQueueIds(videoIds);
     });
@@ -95,11 +110,11 @@ const SearchTab = (props: {socket: any, box: Box}) => {
     }
 
     try {
-      const youtubeSearchResults = await axios.get('https://araza.berrybox.tv/search', {
+      const youtubeSearchRequest = await axios.get(`${Config.API_URL}/search`, {
         params: { value: searchValue },
       });
 
-      setSearchResults(youtubeSearchResults.data);
+      setSearchResults(youtubeSearchRequest.data);
       setSearched(true);
     } catch (error) {
       setError(true);
@@ -107,49 +122,99 @@ const SearchTab = (props: {socket: any, box: Box}) => {
     setSearching(false);
   };
 
-  const submit = async (link: string) => {
-    const submissionPayload: VideoSubmissionRequest = {
-      link,
-      userToken: user._id,
-      boxToken: box._id,
+  const SearchVideo = ({ video, isAlreadyInQueue }) => {
+    const addToQueue = async (flag?: VideoSubmissionRequest['flag']) => {
+      try {
+        await axios.post(`${Config.API_URL}/boxes/${box._id}/queue/video`, {
+          link: video.link,
+          flag,
+        } as Partial<VideoSubmissionRequest>);
+      } catch (error) {
+        setError(true);
+      }
     };
 
-    socket.emit('video', submissionPayload);
-    setSubmitted(true);
+    return (
+      <View style={styles.resultItem}>
+        <View style={{ flex: 0, flexDirection: 'row' }}>
+          <View style={{ paddingRight: 10 }}>
+            <Image
+              style={[{ width: 140, height: 78.75 }, isAlreadyInQueue ? styles.inQueueVideo : null]}
+              source={{ uri: `https://i.ytimg.com/vi/${video.link}/hqdefault.jpg` }}
+            />
+            <DurationIndicator duration={video.duration} withBorder={isAlreadyInQueue} />
+          </View>
+          <View style={{
+            flex: 1,
+            justifyContent: 'center',
+          }}
+          >
+            <Text style={{ color: 'white', fontFamily: 'Monsterrat-Light' }} numberOfLines={3}>
+              {isAlreadyInQueue ? (<Text style={styles.inQueueIndicator}>Already in Queue: </Text>) : null}
+              {video.name}
+            </Text>
+          </View>
+        </View>
+        {!isAlreadyInQueue ? (
+          <View style={{
+            display: 'flex', flexDirection: 'row', alignContent: 'center', justifyContent: 'center', paddingTop: 10,
+          }}
+          >
+            {permissions.includes('addVideo') ? (
+              <Pressable onPress={() => { addToQueue(); }}>
+                <BxButtonComponent options={{
+                  type: 'play',
+                  text: 'Add to Queue',
+                  textDisplay: 'full',
+                }}
+                />
+              </Pressable>
+            ) : null}
+          </View>
+        ) : null}
+        <View style={{
+          display: 'flex', flexDirection: 'row', alignContent: 'center', justifyContent: 'space-between', paddingTop: 10,
+        }}
+        >
+          <>
+            {(permissions.includes('forceNext') && (isAlreadyInQueue || (!isAlreadyInQueue && permissions.includes('addVideo'))))
+                || box.options.berries ? (
+                  <Pressable onPress={() => { addToQueue('next'); }}>
+                    <BxButtonComponent options={{
+                      type: 'forceNext',
+                      text: permissions.includes('forceNext') ? 'Play Next' : 'Play Next - 10 $BC$',
+                      textDisplay: 'full',
+                      context: permissions.includes('forceNext') ? 'primary' : 'berries',
+                    }}
+                    />
+                  </Pressable>
+              ) : null}
+            {(permissions.includes('forcePlay') && (isAlreadyInQueue || (!isAlreadyInQueue && permissions.includes('addVideo'))))
+                || box.options.berries ? (
+                  <Pressable onPress={() => { addToQueue('now'); }}>
+                    <BxButtonComponent options={{
+                      type: 'forcePlay',
+                      text: permissions.includes('forcePlay') ? 'Play Now' : 'Play Now - 50 $BC$',
+                      textDisplay: 'full',
+                      context: permissions.includes('forcePlay') ? 'primary' : 'berries',
+                    }}
+                    />
+                  </Pressable>
+              ) : null}
+          </>
+        </View>
+      </View>
+    );
   };
-
-  const SearchVideo = ({ video, isAlreadyInQueue }) => (
-    <View style={styles.resultItem}>
-      <View>
-        <Image
-          style={[{ width: 140, height: 78.75 }, isAlreadyInQueue ? styles.inQueueVideo : null]}
-          source={{ uri: `https://i.ytimg.com/vi/${video.link}/hqdefault.jpg` }}
-        />
-        <DurationIndicator duration={video.duration} withBorder={isAlreadyInQueue} />
-      </View>
-      <View style={{
-        paddingLeft: 10,
-        width: 200,
-        display: 'flex',
-        justifyContent: 'center',
-      }}
-      >
-        <Text style={{ color: 'white', fontFamily: 'Monsterrat-Light' }} numberOfLines={3}>
-          {isAlreadyInQueue ? (<Text style={styles.inQueueIndicator}>Already in Queue: </Text>) : null}
-          {video.name}
-        </Text>
-      </View>
-    </View>
-  );
 
   const SearchList = () => {
     if (isSearching) {
-      return <ActivityIndicator />;
+      return <BxLoadingIndicator />;
     }
 
     if (youtubeSearchResults.length === 0) {
       if (!hasSearched) {
-        return <></>;
+        return null;
       }
 
       return (
@@ -161,22 +226,9 @@ const SearchTab = (props: {socket: any, box: Box}) => {
       <FlatList
         data={youtubeSearchResults}
         ItemSeparatorComponent={() => <View style={{ backgroundColor: '#3f3f3f', height: 1 }} />}
-        renderItem={(item) => {
-          const video = item.item;
-          const isAlreadyInQueue = videosInQueue.indexOf(video.link) !== -1;
-          if (!isAlreadyInQueue) {
-            return (
-              <TouchableWithoutFeedback
-                onPress={() => submit(video.link)}
-              >
-                <SearchVideo video={video} isAlreadyInQueue={false} />
-              </TouchableWithoutFeedback>
-            );
-          }
-          return (
-            <SearchVideo video={video} isAlreadyInQueue />
-          );
-        }}
+        renderItem={({ item }) => (
+          <SearchVideo video={item} isAlreadyInQueue={videosInQueue.indexOf(item.link) !== -1} />
+        )}
         keyExtractor={(item, index) => index.toString()}
       />
     );
@@ -185,14 +237,21 @@ const SearchTab = (props: {socket: any, box: Box}) => {
   return (
     <View style={styles.container}>
       <View style={styles.searchSpace}>
-        <TextInput
-          style={styles.chatInput}
-          placeholder="Search YouTube for videos to add..."
-          placeholderTextColor="#BBB"
-          onChangeText={(text) => setSearchValue(text)}
-          value={searchValue}
-          onSubmitEditing={() => search()}
-        />
+        {user ? (
+          <View style={{ display: 'flex', flexDirection: 'row' }}>
+            <TextInput
+              style={styles.chatInput}
+              placeholder="Search YouTube for videos to add..."
+              placeholderTextColor="#BBB"
+              onChangeText={(text) => setSearchValue(text)}
+              value={searchValue}
+              onSubmitEditing={() => search()}
+            />
+            {boxOptions?.berries && box?.creator?._id !== user?._id ? (
+              <BerryCounter count={berryCount} />
+            ) : null}
+          </View>
+        ) : null}
         <View style={{ height: '88%' }}>
           <SearchList />
         </View>
