@@ -1,14 +1,20 @@
 import React, { useState, useEffect } from 'react';
 import {
-  KeyboardAvoidingView, StyleSheet, View, Button, Text, TouchableOpacity, ActivityIndicator,
+  KeyboardAvoidingView, StyleSheet, View, Button, Text, TouchableOpacity,
 } from 'react-native';
 import { Switch, Button as IconButton, Snackbar } from 'react-native-paper';
 import axios from 'axios';
 import { Formik } from 'formik';
 import * as yup from 'yup';
 import AsyncStorage from '@react-native-community/async-storage';
+import Config from 'react-native-config';
 
 import FormTextInput from '../components/form-text-input.component';
+import RandomIcon from '../../assets/icons/random-icon.svg';
+import ReplayIcon from '../../assets/icons/replay-icon.svg';
+import BerriesIcon from '../../assets/icons/coin-enabled-icon.svg';
+import LockIcon from '../../assets/icons/lock-icon.svg';
+import BxLoadingIndicator from '../components/bx-loading-indicator.component';
 
 const styles = StyleSheet.create({
   headerContainer: {
@@ -49,6 +55,11 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'space-between',
   },
+  modeDefinition: {
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
   modeTitle: {
     fontSize: 20,
     fontFamily: 'Montserrat-SemiBold',
@@ -79,14 +90,15 @@ const CreateBoxScreen = ({ navigation }) => {
     getSession();
   }, []);
 
-  const createBox = async (boxInputData: { name: string, options: BoxOptions}) => {
+  const createBox = async (boxInputData: { name: string, private: boolean, options: BoxOptions}) => {
     setCreating(true);
-    const box = await axios.post('https://araza.berrybox.tv/boxes', {
+    const box = await axios.post(`${Config.API_URL}/boxes`, {
       creator: user._id,
       name: boxInputData.name,
       description: null,
       lang: 'en',
       open: true,
+      private: boxInputData.private,
       options: boxInputData.options,
     });
     setBox(box.data);
@@ -112,13 +124,15 @@ const CreateBoxScreen = ({ navigation }) => {
       >
         <Formik
           initialValues={{
-            name: '', random: false, loop: false, berries: true,
+            name: '', private: false, random: false, loop: false, berries: true,
           }}
           validationSchema={
               yup.object().shape({
                 name: yup
                   .string()
                   .required('The box name is required'),
+                private: yup
+                  .boolean(),
                 random: yup
                   .boolean(),
                 loop: yup
@@ -129,6 +143,7 @@ const CreateBoxScreen = ({ navigation }) => {
             }
           onSubmit={(values) => createBox({
             name: values.name,
+            private: values.private,
             options: { random: values.random, loop: values.loop, berries: values.berries },
           })}
         >
@@ -149,7 +164,28 @@ const CreateBoxScreen = ({ navigation }) => {
               </View>
               <View style={styles.modeContainer}>
                 <View style={styles.modeSpace}>
-                  <Text style={styles.modeTitle}>Random</Text>
+                  <View style={styles.modeDefinition}>
+                    <View style={{ paddingRight: 5 }}>
+                      <LockIcon width={20} height={20} fill="white" />
+                    </View>
+                    <Text style={styles.modeTitle}>Access Restriction</Text>
+                  </View>
+                  <Switch
+                    value={values.private}
+                    onValueChange={(value) => setFieldValue('private', value)}
+                    color="#009AEB"
+                  />
+                </View>
+                <Text style={styles.modeHelper}>Your box will not appear in the home page. You will only be able to grant access by sharing its link directly.</Text>
+              </View>
+              <View style={styles.modeContainer}>
+                <View style={styles.modeSpace}>
+                  <View style={styles.modeDefinition}>
+                    <View style={{ paddingRight: 5 }}>
+                      <RandomIcon width={20} height={20} fill="white" />
+                    </View>
+                    <Text style={styles.modeTitle}>Pick Videos at Random</Text>
+                  </View>
                   <Switch
                     value={values.random}
                     onValueChange={(value) => setFieldValue('random', value)}
@@ -162,7 +198,12 @@ const CreateBoxScreen = ({ navigation }) => {
               </View>
               <View style={styles.modeContainer}>
                 <View style={styles.modeSpace}>
-                  <Text style={styles.modeTitle}>Loop</Text>
+                  <View style={styles.modeDefinition}>
+                    <View style={{ paddingRight: 5 }}>
+                      <ReplayIcon width={20} height={20} fill="white" />
+                    </View>
+                    <Text style={styles.modeTitle}>Loop Queue</Text>
+                  </View>
                   <Switch
                     value={values.loop}
                     onValueChange={(value) => setFieldValue('loop', value)}
@@ -173,7 +214,12 @@ const CreateBoxScreen = ({ navigation }) => {
               </View>
               <View style={styles.modeContainer}>
                 <View style={styles.modeSpace}>
-                  <Text style={styles.modeTitle}>Berries</Text>
+                  <View style={styles.modeDefinition}>
+                    <View style={{ paddingRight: 5 }}>
+                      <BerriesIcon width={20} height={20} fill="white" />
+                    </View>
+                    <Text style={styles.modeTitle}>Berries System</Text>
+                  </View>
                   <Switch
                     value={values.berries}
                     onValueChange={(value) => setFieldValue('berries', value)}
@@ -192,7 +238,7 @@ const CreateBoxScreen = ({ navigation }) => {
                   onPress={handleSubmit}
                 />
               ) : (
-                <ActivityIndicator />
+                <BxLoadingIndicator />
               )}
               <Snackbar
                 visible={box}
